requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
psycopg[binary]==3.1.12
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0